fix CodeQL errors in XRT drivers (#6127)

+++ src/runtime_src/core/include/xgq_impl.h
@@ -289,13 +289,13 @@ static inline void xgq_ring_write_consumed(uint64_t io_hdl, struct xgq_ring *rin
 static inline uint64_t xgq_ring_slot_ptr_produced(struct xgq_ring *ring)
 {
 	return ring->xr_slot_addr +
-	       ring->xr_slot_sz * (ring->xr_produced & (ring->xr_slot_num - 1));
+	       (uint64_t)ring->xr_slot_sz * (ring->xr_produced & (ring->xr_slot_num - 1));
 }
 
 static inline uint64_t xgq_ring_slot_ptr_consumed(struct xgq_ring *ring)
 {
 	return ring->xr_slot_addr +
-	       ring->xr_slot_sz * (ring->xr_consumed & (ring->xr_slot_num - 1));
+	       (uint64_t)ring->xr_slot_sz * (ring->xr_consumed & (ring->xr_slot_num - 1));
 }
 
 static inline int xgq_can_produce(struct xgq *xgq)
